Break out the `policy ...` command hierarchy
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`policy ...` hierarchy.
