+++ pkg/cmd/pulumi/policy/policy_disable.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2020, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,12 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package policy
 
 import (
-	"encoding/json"
-	"fmt"
-
 	"github.com/pulumi/pulumi/pkg/v3/backend"
 	cmdBackend "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/backend"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
@@ -25,51 +22,43 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func newPolicyValidateCmd() *cobra.Command {
-	var argConfig string
+type policyDisableArgs struct {
+	policyGroup string
+	version     string
+}
+
+func newPolicyDisableCmd() *cobra.Command {
+	args := policyDisableArgs{}
 
 	cmd := &cobra.Command{
-		Use:   "validate-config <org-name>/<policy-pack-name> <version>",
-		Args:  cmdutil.ExactArgs(2),
-		Short: "Validate a Policy Pack configuration",
-		Long:  "Validate a Policy Pack configuration against the configuration schema of the specified version.",
+		Use:   "disable <org-name>/<policy-pack-name>",
+		Args:  cmdutil.ExactArgs(1),
+		Short: "Disable a Policy Pack for a Pulumi organization",
+		Long:  "Disable a Policy Pack for a Pulumi organization",
 		Run: cmd.RunCmdFunc(func(cmd *cobra.Command, cliArgs []string) error {
 			ctx := cmd.Context()
 			// Obtain current PolicyPack, tied to the Pulumi Cloud backend.
+			var err error
 			policyPack, err := requirePolicyPack(ctx, cliArgs[0], cmdBackend.DefaultLoginManager)
 			if err != nil {
 				return err
 			}
 
-			// Get version from cmd argument
-			version := &cliArgs[1]
-
-			// Load the configuration from the user-specified JSON file into config object.
-			var config map[string]*json.RawMessage
-			if argConfig != "" {
-				config, err = loadPolicyConfigFromFile(argConfig)
-				if err != nil {
-					return err
-				}
-			}
-
-			err = policyPack.Validate(ctx,
-				backend.PolicyPackOperation{
-					VersionTag: version,
-					Scopes:     backend.CancellationScopes,
-					Config:     config,
-				})
-			if err != nil {
-				return err
-			}
-			fmt.Println("Policy Pack configuration is valid.")
-			return nil
+			// Attempt to disable the Policy Pack.
+			return policyPack.Disable(ctx, args.policyGroup, backend.PolicyPackOperation{
+				VersionTag: &args.version, Scopes: backend.CancellationScopes,
+			})
 		}),
 	}
 
-	cmd.Flags().StringVar(&argConfig, "config", "",
-		"The file path for the Policy Pack configuration file")
-	cmd.MarkFlagRequired("config") //nolint:errcheck
+	cmd.PersistentFlags().StringVar(
+		&args.policyGroup, "policy-group", "",
+		"The Policy Group for which the Policy Pack will be disabled; if not specified, the default Policy Group is used")
+
+	cmd.PersistentFlags().StringVar(
+		&args.version, "version", "",
+		"The version of the Policy Pack that will be disabled; "+
+			"if not specified, any enabled version of the Policy Pack will be disabled")
 
 	return cmd
 }
