+++ pkg/cmd/pulumi/policy/policy_group_ls.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package policy
 
 import (
 	"errors"
