+++ pkg/cmd/pulumi/policy/policy_publish.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2018, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package policy
 
 import (
 	"context"
@@ -108,7 +108,7 @@ func (cmd *policyPublishCmd) Run(ctx context.Context, lm cmdBackend.LoginManager
 		return err
 	}
 
-	proj, _, root, err := readPolicyProject(pwd)
+	proj, _, root, err := ReadPolicyProject(pwd)
 	if err != nil {
 		return err
 	}
