+++ pkg/cmd/pulumi/policy/testdata/PulumiPolicy.yaml
