+++ pkg/cmd/pulumi/util.go
@@ -15,20 +15,13 @@
 package main
 
 import (
-	"context"
 	"errors"
 	"fmt"
-	"path/filepath"
 	"strconv"
-
-	"github.com/opentracing/opentracing-go"
 
 	"github.com/pulumi/pulumi/pkg/v3/backend"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/newcmd"
-	"github.com/pulumi/pulumi/pkg/v3/engine"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/plugin"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 )
 
@@ -65,26 +58,6 @@ func readProjectForUpdate(ws pkgWorkspace.Context, clientAddress string) (*works
 		})
 	}
 	return proj, root, nil
-}
-
-// readPolicyProject attempts to detect and read a Pulumi PolicyPack project for the current
-// workspace. If the project is successfully detected and read, it is returned along with the path
-// to its containing directory, which will be used as the root of the project's Pulumi program.
-func readPolicyProject(pwd string) (*workspace.PolicyPackProject, string, string, error) {
-	// Now that we got here, we have a path, so we will try to load it.
-	path, err := workspace.DetectPolicyPackPathFrom(pwd)
-	if err != nil {
-		return nil, "", "", fmt.Errorf("failed to find current Pulumi project because of "+
-			"an error when searching for the PulumiPolicy.yaml file (searching upwards from %s)"+": %w", pwd, err)
-	} else if path == "" {
-		return nil, "", "", fmt.Errorf("no PulumiPolicy.yaml project file found (searching upwards from %s)", pwd)
-	}
-	proj, err := workspace.LoadPolicyPack(path)
-	if err != nil {
-		return nil, "", "", fmt.Errorf("failed to load Pulumi policy project located at %q: %w", path, err)
-	}
-
-	return proj, path, filepath.Dir(path), nil
 }
 
 // updateFlagsToOptions ensures that the given update flags represent a valid combination.  If so, an UpdateOptions
@@ -130,41 +103,4 @@ func getRefreshOption(proj *workspace.Project, refresh string) (bool, error) {
 
 	// the default functionality right now is to always skip a refresh
 	return false, nil
-}
-
-func installPolicyPackDependencies(ctx context.Context, root string, proj *workspace.PolicyPackProject) error {
-	span := opentracing.SpanFromContext(ctx)
-	// Bit of a hack here. Creating a plugin context requires a "program project", but we've only got a
-	// policy project. Ideally we should be able to make a plugin context without any related project. But
-	// fow now this works.
-	projinfo := &engine.Projinfo{Proj: &workspace.Project{
-		Main:    proj.Main,
-		Runtime: proj.Runtime,
-	}, Root: root}
-	_, main, pluginCtx, err := engine.ProjectInfoContext(
-		projinfo,
-		nil,
-		cmdutil.Diag(),
-		cmdutil.Diag(),
-		nil,
-		false,
-		span,
-		nil,
-	)
-	if err != nil {
-		return err
-	}
-	defer pluginCtx.Close()
-
-	programInfo := plugin.NewProgramInfo(pluginCtx.Root, pluginCtx.Pwd, main, proj.Runtime.Options())
-	lang, err := pluginCtx.Host.LanguageRuntime(proj.Runtime.Name(), programInfo)
-	if err != nil {
-		return fmt.Errorf("failed to load language plugin %s: %w", proj.Runtime.Name(), err)
-	}
-
-	if err = lang.InstallDependencies(plugin.InstallDependenciesRequest{Info: programInfo}); err != nil {
-		return fmt.Errorf("installing dependencies failed: %w", err)
-	}
-
-	return nil
 }
